Few more cleanups.

+++ clients/generator.c
@@ -21,7 +21,7 @@ static void get_random_string(char *buffer, size_t size)
 
 void pairs_free(pairs_st *pairs)
 {
-  unsigned int x;
+  uint32_t x;
 
   if (!pairs)
     return;
@@ -36,7 +36,7 @@ void pairs_free(pairs_st *pairs)
   free(pairs);
 }
 
-pairs_st *pairs_generate(unsigned long long number_of, size_t value_length)
+pairs_st *pairs_generate(uint64_t number_of, size_t value_length)
 {
   unsigned int x;
   pairs_st *pairs;
